# Optimize TypedValue.Compare for value reflector