--- a/value/map.go
+++ b/value/map.go
@@ -18,7 +18,6 @@ package value
 
 import (
 	"sort"
-	"strings"
 )
 
 // Map represents a Map or go structure.
@@ -40,6 +39,139 @@ type Map interface {
 	Iterate(func(key string, value Value) bool) bool
 	// Length returns the number of items in the map.
 	Length() int
+	// Empty returns true if the map is empty.
+	Empty() bool
+	// Zip iterates over the entries of two maps together. If both maps contain a value for a given key, fn is called
+	// with the values from both maps, otherwise it is called with the value of the map that contains the key and nil
+	// for the map that does not contain the key. Returning false in the closure prematurely stops the iteration.
+	Zip(other Map, order MapTraverseOrder, fn func(key string, lhs, rhs Value) bool) bool
+	// Recycle gives back this Map once it is no longer needed. The
+	// value shouldn't be used after this call.
+	Recycle()
+}
+
+// MapTraverseOrder defines the map traversal ordering available.
+type MapTraverseOrder int
+
+const (
+	// Unordered indicates that the map traversal has no ordering requirement.
+	Unordered = iota
+	// LexicalKeyOrder indicates that the map traversal is ordered by key, lexically.
+	LexicalKeyOrder
+)
+
+// MapZip iterates over the entries of two maps together. If both maps contain a value for a given key, fn is called
+// with the values from both maps, otherwise it is called with the value of the map that contains the key and nil
+// for the other map. Returning false in the closure prematurely stops the iteration.
+func MapZip(lhs, rhs Map, order MapTraverseOrder, fn func(key string, lhs, rhs Value) bool) bool {
+	if lhs != nil {
+		return lhs.Zip(rhs, order, fn)
+	}
+	if rhs != nil {
+		return rhs.Zip(lhs, order, func(key string, rhs, lhs Value) bool { // arg positions of lhs and rhs deliberately swapped
+			return fn(key, lhs, rhs)
+		})
+	}
+	return true
+}
+
+// defaultMapZip provides a default implementation of Zip for implementations that do not need to provide
+// their own optimized implementation.
+func defaultMapZip(lhs, rhs Map, order MapTraverseOrder, fn func(key string, lhs, rhs Value) bool) bool {
+	switch order {
+	case Unordered:
+		return unorderedMapZip(lhs, rhs, fn)
+	case LexicalKeyOrder:
+		return lexicalKeyOrderedMapZip(lhs, rhs, fn)
+	default:
+		panic("Unsupported map order")
+	}
+}
+
+func unorderedMapZip(lhs, rhs Map, fn func(key string, lhs, rhs Value) bool) bool {
+	if (lhs == nil || lhs.Empty()) && (rhs == nil || rhs.Empty()) {
+		return true
+	}
+
+	if lhs != nil {
+		ok := lhs.Iterate(func(key string, lhsValue Value) bool {
+			var rhsValue Value
+			if rhs != nil {
+				if item, ok := rhs.Get(key); ok {
+					rhsValue = item
+					defer rhsValue.Recycle()
+				}
+			}
+			return fn(key, lhsValue, rhsValue)
+		})
+		if !ok {
+			return false
+		}
+	}
+	if rhs != nil {
+		return rhs.Iterate(func(key string, rhsValue Value) bool {
+			if lhs == nil || !lhs.Has(key) {
+				return fn(key, nil, rhsValue)
+			}
+			return true
+		})
+	}
+	return true
+}
+
+func lexicalKeyOrderedMapZip(lhs, rhs Map, fn func(key string, lhs, rhs Value) bool) bool {
+	var lhsLength, rhsLength int
+	var orderedLength int // rough estimate of length of union of map keys
+	if lhs != nil {
+		lhsLength = lhs.Length()
+		orderedLength = lhsLength
+	}
+	if rhs != nil {
+		rhsLength = rhs.Length()
+		if rhsLength > orderedLength {
+			orderedLength = rhsLength
+		}
+	}
+	if lhsLength == 0 && rhsLength == 0 {
+		return true
+	}
+
+	ordered := make([]string, 0, orderedLength)
+	if lhs != nil {
+		lhs.Iterate(func(key string, _ Value) bool {
+			ordered = append(ordered, key)
+			return true
+		})
+	}
+	if rhs != nil {
+		rhs.Iterate(func(key string, _ Value) bool {
+			if lhs == nil || !lhs.Has(key) {
+				ordered = append(ordered, key)
+			}
+			return true
+		})
+	}
+	sort.Strings(ordered)
+	for _, key := range ordered {
+		var litem, ritem Value
+		if lhs != nil {
+			litem, _ = lhs.Get(key)
+		}
+		if rhs != nil {
+			ritem, _ = rhs.Get(key)
+		}
+		ok := fn(key, litem, ritem)
+		if litem != nil {
+			litem.Recycle()
+		}
+		if ritem != nil {
+			ritem.Recycle()
+		}
+		if !ok {
+			return false
+		}
+	}
+	return true
 }
 
 // MapLess compares two maps lexically.
@@ -49,68 +181,55 @@ func MapLess(lhs, rhs Map) bool {
 
 // MapCompare compares two maps lexically.
 func MapCompare(lhs, rhs Map) int {
-	lorder := make([]string, 0, lhs.Length())
-	lhs.Iterate(func(key string, _ Value) bool {
-		lorder = append(lorder, key)
-		return true
-	})
-	sort.Strings(lorder)
-	rorder := make([]string, 0, rhs.Length())
-	rhs.Iterate(func(key string, _ Value) bool {
-		rorder = append(rorder, key)
-		return true
-	})
-	sort.Strings(rorder)
-
-	i := 0
-	for {
-		if i >= len(lorder) && i >= len(rorder) {
-			// Maps are the same length and all items are equal.
-			return 0
-		}
-		if i >= len(lorder) {
-			// LHS is shorter.
-			return -1
-		}
-		if i >= len(rorder) {
-			// RHS is shorter.
-			return 1
-		}
-		if c := strings.Compare(lorder[i], rorder[i]); c != 0 {
-			return c
-		}
-		litem, _ := lhs.Get(lorder[i])
-		ritem, _ := rhs.Get(rorder[i])
-		c := Compare(litem, ritem)
-		if litem != nil {
-			litem.Recycle()
-		}
-		if ritem != nil {
-			ritem.Recycle()
-		}
-		if c != 0 {
-			return c
-		}
-		// The items are equal; continue.
-		i++
+	c := 0
+	var llength, rlength int
+	if lhs != nil {
+		llength = lhs.Length()
 	}
+	if rhs != nil {
+		rlength = rhs.Length()
+	}
+	if llength == 0 && rlength == 0 {
+		return 0
+	}
+	i := 0
+	MapZip(lhs, rhs, LexicalKeyOrder, func(key string, lhs, rhs Value) bool {
+		switch {
+		case i == llength:
+			c = -1
+		case i == rlength:
+			c = 1
+		case lhs == nil:
+			c = 1
+		case rhs == nil:
+			c = -1
+		default:
+			c = Compare(lhs, rhs)
+		}
+		i++
+		return c == 0
+	})
+	return c
 }
 
 // MapEquals returns true if lhs == rhs, false otherwise. This function
 // acts on generic types and should not be used by callers, but can help
 // implement Map.Equals.
-// WARN: This is a naive implementation, calling lhs.Equals(rhs) is typically far more efficient.
+// WARN: This is a naive implementation, calling lhs.Equals(rhs) is typically the most efficient.
 func MapEquals(lhs, rhs Map) bool {
+	if lhs == nil && rhs == nil {
+		return true
+	}
+	if lhs == nil || rhs == nil {
+		return false
+	}
 	if lhs.Length() != rhs.Length() {
 		return false
 	}
-	return lhs.Iterate(func(k string, v Value) bool {
-		vo, ok := rhs.Get(k)
-		if !ok {
+	return MapZip(lhs, rhs, Unordered, func(key string, lhs, rhs Value) bool {
+		if lhs == nil || rhs == nil {
 			return false
 		}
-		equal := Equals(v, vo)
-		vo.Recycle()
-		return equal
+		return Equals(lhs, rhs)
 	})
 }