--- a/value/mapreflect.go
+++ b/value/mapreflect.go
@@ -29,6 +29,11 @@ func (r mapReflect) Length() int {
 	return val.Len()
 }
 
+func (r mapReflect) Empty() bool {
+	val := r.Value
+	return val.Len() == 0
+}
+
 func (r mapReflect) Get(key string) (Value, bool) {
 	k, v, ok := r.get(key)
 	if !ok {
@@ -74,21 +79,17 @@ func (r mapReflect) Iterate(fn func(string, Value) bool) bool {
 	vr := reflectPool.Get().(*valueReflect)
 	defer vr.Recycle()
 	return eachMapEntry(r.Value, func(e *TypeReflectCacheEntry, key reflect.Value, value reflect.Value) bool {
-		// TODO: Track cache entry
-		return fn(key.String(), vr.mustReuse(value, nil, &r.Value, &key))
+		return fn(key.String(), vr.mustReuse(value, e, &r.Value, &key))
 	})
 }
 
 func eachMapEntry(val reflect.Value, fn func(*TypeReflectCacheEntry, reflect.Value, reflect.Value) bool) bool {
 	iter := val.MapRange()
-	var entry *TypeReflectCacheEntry
+	entry := TypeReflectEntryOf(val.Type().Elem())
 	for iter.Next() {
 		next := iter.Value()
 		if !next.IsValid() {
 			continue
-		}
-		if entry == nil {
-			entry = TypeReflectEntryOf(next.Type())
 		}
 		if !fn(entry, iter.Key(), next) {
 			return false
@@ -117,12 +118,80 @@ func (r mapReflect) Equals(m Map) bool {
 	}
 	vr := reflectPool.Get().(*valueReflect)
 	defer vr.Recycle()
+	entry := TypeReflectEntryOf(r.Value.Type().Elem())
 	return m.Iterate(func(key string, value Value) bool {
 		_, lhsVal, ok := r.get(key)
 		if !ok {
 			return false
 		}
-		// TODO: Track cache entry
-		return Equals(vr.mustReuse(lhsVal, nil, nil, nil), value)
+		return Equals(vr.mustReuse(lhsVal, entry, nil, nil), value)
 	})
+}
+
+func (r mapReflect) Zip(other Map, order MapTraverseOrder, fn func(key string, lhs, rhs Value) bool) bool {
+	if otherMapReflect, ok := other.(*mapReflect); ok && order == Unordered {
+		return r.unorderedReflectZip(otherMapReflect, fn)
+	}
+	return defaultMapZip(&r, other, order, fn)
+}
+
+// unorderedReflectZip provides an optimized unordered zip for mapReflect types.
+func (r mapReflect) unorderedReflectZip(other *mapReflect, fn func(key string, lhs, rhs Value) bool) bool {
+	if r.Empty() && (other == nil || other.Empty()) {
+		return true
+	}
+
+	lhs := r.Value
+	lhsEntry := TypeReflectEntryOf(lhs.Type().Elem())
+
+	vl := reflectPool.Get().(*valueReflect)
+	defer vl.Recycle()
+	vr := reflectPool.Get().(*valueReflect)
+	defer vr.Recycle()
+
+	// map lookup via reflection is expensive enough that it is better to keep track of visited keys
+	visited := map[string]struct{}{}
+
+	if other != nil {
+		rhs := other.Value
+		rhsEntry := TypeReflectEntryOf(rhs.Type().Elem())
+		iter := rhs.MapRange()
+		for iter.Next() {
+			key := iter.Key()
+			keyString := key.String()
+			next := iter.Value()
+			if !next.IsValid() {
+				continue
+			}
+			rhsVal := vr.mustReuse(next, rhsEntry, &rhs, &key)
+			visited[keyString] = struct{}{}
+			var lhsVal Value
+			if _, v, ok := r.get(keyString); ok {
+				lhsVal = vl.mustReuse(v, lhsEntry, &lhs, &key)
+			}
+			if !fn(keyString, lhsVal, rhsVal) {
+				return false
+			}
+		}
+	}
+
+	iter := lhs.MapRange()
+	for iter.Next() {
+		key := iter.Key()
+		if _, ok := visited[key.String()]; ok {
+			continue
+		}
+		next := iter.Value()
+		if !next.IsValid() {
+			continue
+		}
+		if !fn(key.String(), vl.mustReuse(next, lhsEntry, &lhs, &key), nil) {
+			return false
+		}
+	}
+	return true
+}
+
+func (r *mapReflect) Recycle() {
+	mapReflectPool.Put(r)
 }