… value/structreflect.go
… value/valuereflect.go
… value/valuereflect_test.go